requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.8.2
python-dotenv>=1.0.0
httpx>=0.27.0
mistralai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-mistralai>=0.1.0
elevenlabs>=0.2.0
openai-whisper>=20240314
torch>=2.0.0
